if fi

+++ install_easystroke.sh
@@ -6,14 +6,24 @@ sudo apt-get install -y xclip easystroke
   
 cd ~ #TODO verifier que .easystroke n'existe pas
 
-rm -rf .easystroke
+if [  -d .easystroke ]; then
+	rm -rf .easystroke
+fi
 
 echo "Retrieve easystroke config"
 git clone https://gitlab.com/Sraleik/easystrokeConfig.git .easystroke
 echo "easystroke config -> in ~/.easystroke"
 
 echo "easystroke launch at start"
-rm -f ~/.config/autostart/easystroke.desktop
+
+if [ ! -d ~/.config/autostart ]; then
+	mkdir ~/.config/autostart
+fi
+
+if [ -f ~/.config/autostart/easystroke.desktop ]; then
+	rm -f ~/.config/autostart/easystroke.desktop
+fi
+
 touch ~/.config/autostart/easystroke.desktop
 
 cat << EOF >> ~/.config/autostart/easystroke.desktop
